+++ assets/css/app.scss
@@ -57,3 +57,38 @@ input, button, select {
   margin-right: auto;
   margin-left: auto;
 }
+.tabla {
+  box-shadow: #00000024 0px 1px 10px 0px;
+  border-radius: 10px;
+  font-size: 0.8rem;
+  h2 {
+    padding: .5rem;
+    background: #3f51b5;
+    border-top-left-radius: 10px;
+    color: white;
+    border-top-right-radius: 10px;
+    font-size: 1.3rem;
+    font-weight: bold;
+  }
+  .cuerpo-tabla {
+    margin-left: 20px;
+    margin-right: 20px;
+  }
+  table {
+    margin: 20px auto;
+    border-collapse: collapse;
+    border-radius: 10px;
+    overflow: hidden;
+    box-shadow: #00000024 0px 1px 10px 0px;
+    border: none;
+    .header th {
+      background: #dee2e6;
+      min-width: 3rem;
+      border: none;
+      padding: 0.2rem 0.5rem;
+    }
+    .second th{
+      background: darken(#dee2e6, 10%);
+    }
+  }
+}
