+++ src/Service/SimuladorService.php
@@ -9,6 +9,15 @@ use App\Entity\Simulador;
 
 class SimuladorService
 {
+    private $planificacionService;
+    private $intercambioService;
+
+    public function __construct(PlanificacionService $planificacionService, IntercambioService $intercambioService)
+    {
+        $this->planificacionService = $planificacionService;
+        $this->intercambioService = $intercambioService;
+    }
+
     function simular(Simulador $simulador)
     {
         $memoria = $simulador->getMemoria();
@@ -38,7 +47,7 @@ class SimuladorService
                     case 'ff':
                         //Llamo a la funcion de asignación de memoria
                         list($cola_listos, $cola_nuevos, $particiones) =
-                            $this->asignacionParticionesFijasFF($cola_listos, $cola_nuevos, $particiones)
+                            $this->intercambioService->asignacionParticionesFijasFF($cola_listos, $cola_nuevos, $particiones)
                         ;
                         break;
                     case 'bf':
@@ -61,8 +70,6 @@ class SimuladorService
 //                        dd('no hay PARTICIONES VARIABLES AUN');
 //                        break;
 //                }
-            } else {
-                dd('Se rompio todo');
             }
 
             if ($t == 0) {
@@ -89,7 +96,7 @@ class SimuladorService
              * recorriendo los procesos que se bloquearon por una entrada/salida
              */
             list($cola_bloqueados, $cola_nuevos) =
-              $this->tratarBloqueados($cola_bloqueados, $cola_nuevos)
+              $this->planificacionService->tratarBloqueados($cola_bloqueados, $cola_nuevos)
             ;
 
             switch ($simulador->getAlgoritmoPlanificacion()) {
@@ -100,7 +107,7 @@ class SimuladorService
                      * la rafaga actual
                      */
                     list($cola_listos, $cola_bloqueados, $particiones, $rafagaActual) =
-                        $this->fcfs($cola_listos, $cola_bloqueados, $particiones, $rafagaActual)
+                        $this->planificacionService->fcfs($cola_listos, $cola_bloqueados, $particiones, $rafagaActual)
                     ;
                     break;
                 case 'rr':
@@ -169,29 +176,6 @@ class SimuladorService
         ];
     }
 
-    function asignacionParticionesFijasFF($cola_listos, $cola_nuevos, $particiones)
-    {
-        //Recorro las particiones
-        foreach ($particiones as $particionKey => $particion) {
-            //Recorro los procesos
-            foreach ($cola_nuevos as $procesoKey => $proceso) {
-                //Asigno si el proceso cabe en la particion y si tiene de status nuevo
-                if ($particiones[$particionKey]['proceso_asignado'] == null and
-                  $proceso['size'] <= $particion['size']
-                ) {
-                    //Asigno el proceso a la partición
-                    $particiones[$particionKey]['proceso_asignado'] = $cola_nuevos[$procesoKey];
-                    //Pongo el proceso en la cola de listos
-                    array_push($cola_listos, $cola_nuevos[$procesoKey]);
-                    //Saco el proceso de la cola de nuevos
-                    unset($cola_nuevos[$procesoKey]);
-                }
-            }
-        }
-
-        return [$cola_listos, $cola_nuevos, $particiones];
-    }
-
     function guardarProcesosColaNuevos($cola_nuevos, $procesos, $rafaga)
     {
         foreach ($procesos as $key => $proceso) {
@@ -203,80 +187,6 @@ class SimuladorService
         }
 
         return $cola_nuevos;
-    }
-
-    function liberarProcesoDeMemoria($proceso, $particiones){
-        foreach ($particiones as $key => $particion) {
-            if ($particion['proceso_asignado']['id'] == $proceso['id'] ) {
-                $particiones[$key]['proceso_asignado'] = null;
-            }
-        }
-        return $particiones;
-    }
-
-    function fcfs($cola_listos, $cola_bloqueados, $particiones, $rafagaActual) {
-        if (!empty($cola_listos)) {
-            $procesoEnTratamiento = $cola_listos[0];
-            $ciclo = $procesoEnTratamiento['ciclo'];
-            $rafagaActual['ejecuto'] = $procesoEnTratamiento; //Cargar proceso ejecutado
-
-            if ($ciclo[0]['tipo'] == 'irrupcion') {
-                $tiempo_remanente = $ciclo[0]['valor'] - 1;
-
-                if ($tiempo_remanente == 0 && isset($ciclo[1])) { //Si se termina la irrupcion y viene un bloqueo
-
-                    unset($ciclo[0]); //Sacar la irrupción que llego a cero del ciclo
-                    unset($cola_listos[0]); //Sacar el proceso de la cola de listos
-                    $procesoEnTratamiento['ciclo'] = array_values($ciclo); //Actualizar el proceso sin la irrupción que termino
-                    $particiones = $this->liberarProcesoDeMemoria($procesoEnTratamiento, $particiones); //Libero la memoria
-                    array_push($cola_bloqueados, $procesoEnTratamiento);
-
-                    $rafagaActual['bloqueo'] = $procesoEnTratamiento; //Cargar proceso ejecutado
-
-                } else if ($tiempo_remanente == 0 && !isset($ciclo[1])) { // Si termina la irrupción y termina el proceso
-
-                    unset($ciclo[0]); //Sacar la irrupción que llego a cero del ciclo
-                    unset($cola_listos[0]); //Sacar el proceso de la cola de listos
-                    $procesoEnTratamiento['ciclo'] = array_values($ciclo); //Actualizar el ciclo del proceso
-                    $particiones = $this->liberarProcesoDeMemoria($procesoEnTratamiento, $particiones); //Libero la memoria
-                    $rafagaActual['finalizo'] = $procesoEnTratamiento; //Cargar proceso finalizado
-
-                } else {
-                    //El proceso se ejecuta normalmente y sigue en CPU
-                    $ciclo[0]['valor'] = $tiempo_remanente; //Se resta la irrupcion
-                    $cola_listos[0]['ciclo'] = $ciclo; //Se actualiza el ciclo en la cola de listos
-                }
-            }
-        }
-
-        return [array_values($cola_listos), array_values($cola_bloqueados), $particiones, $rafagaActual];
-    }
-
-
-    function tratarBloqueados($cola_bloqueados, $cola_nuevos) {
-        if (!empty($cola_bloqueados)) {
-            $procesoEnTratamiento = $cola_bloqueados[0];
-            $ciclo = $cola_bloqueados[0]['ciclo'];
-
-            if ($ciclo[0]['tipo'] == 'bloqueo') {
-                $bloqueo_remanente = $ciclo[0]['valor'] - 1;
-
-                if ($bloqueo_remanente == 0 ) { //Si se termina la irrupcion y viene un bloqueo
-                    unset($ciclo[0]); //Sacar la irrupción que llego a cero del ciclo
-                    unset($cola_bloqueados[0]); //Sacar el proceso de la cola de listos
-                    $procesoEnTratamiento['ciclo'] = array_values($ciclo); //Actualizar el proceso sin la irrupción que termino
-                    array_push($cola_nuevos, $procesoEnTratamiento); //El proceso vuelve a la cola de nuevos a competir por memoria
-
-                } else {
-                    //El proceso se ejecuta normalmente y sigue en E/S
-                    $ciclo[0]['valor'] = $bloqueo_remanente; //Se resta la irrupcion
-                    $cola_bloqueados[0]['ciclo'] = $ciclo; //Se actualiza el ciclo en la cola de bloqueados
-
-                }
-            }
-        }
-
-        return [array_values($cola_bloqueados), array_values($cola_nuevos)];
     }
 
     function finalizoSimulador($cola_listos, $cola_bloqueados, $cola_nuevos) {
