+++ templates/home/index.html.twig
@@ -19,6 +19,7 @@
                 Para iniciar una simulación haga click en el siguiente botón:
             </p>
             <a class="btn btn-success" href="{{ path('new') }}">Comenzar simulación</a>
+            <a class="btn btn-success" href="{{ path('simulador_index') }}">Listado de simulaciones</a>
         </div>
     </div>
 {% endblock %}
