+++ templates/simulador/new.scss
@@ -220,6 +220,7 @@ form, .list, .particion-list {
     overflow: hidden;
     box-shadow: #00000024 0px 1px 10px 0px;
     border: none;
+    margin: auto;
     .header {
       th {
         background: #dee2e6;
